tests/wpt/mozilla/tests/mozilla/window_performance_topLevelDomComplete.html

<html>
<head>
  <title>Performance toLevelDomComplete</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>
async_test(function(t) {
    window.onload = t.step_func(function () {
    var entries = window.performance.getEntries();
      assert_greater_than(entries.length, 0, "There should be at least one navigation timing instance");

      // The measurements are sometimes a few ns apart
      assert_true(Math.abs(entries[0].domComplete - entries[0].topLevelDomComplete) < 20);
      t.done()
    }
  );
}, "performance.topLevelDomComplete");
</script>
</body>
</html>
